backend/requirements.txt
azure-functions
qrcode[pil]
